implements modularization
