+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,5 @@
 <resources>
     <string name="app_name">Tourism App</string>
-    <string name="image_description">Image</string>
-    <string name="text_primary">Sample of title text</string>
-    <string name="text_subtitle">Sample of subtitle text</string>
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
