+++ build.gradle
@@ -4,7 +4,7 @@ buildscript {
     ext.kotlin_version = '1.4.10'
     ext.appcompat_version = '1.2.0'
     ext.core_ktx_version = "1.3.2"
-    ext.constraint_version = "2.0.2"
+    ext.constraint_version = "2.0.4"
     ext.legacy_support_version = "1.0.0"
     ext.junit_version = '4.13.1'
     ext.androidx_junit_version = "1.1.2"
