+++ core/src/main/java/com/ryanrvldo/tourismapp/core/data/NetworkBoundResource.kt
@@ -14,16 +14,32 @@ abstract class NetworkBoundResource<ResultType, RequestType>(private val mExecut
             when (val apiResponse = createCall().first()) {
                 is ApiResponse.Success -> {
                     saveCallResult(apiResponse.data)
-                    emitAll(loadFromDB().map { Resource.Success(it) })
+                    emitAll(loadFromDB().map {
+                        Resource.Success(
+                            it
+                        )
+                    })
                 }
-                is ApiResponse.Empty -> emitAll(loadFromDB().map { Resource.Success(it) })
+                is ApiResponse.Empty -> emitAll(loadFromDB().map {
+                    Resource.Success(
+                        it
+                    )
+                })
                 is ApiResponse.Error -> {
                     onFetchFailed()
-                    emit(Resource.Error<ResultType>(apiResponse.errorMessage))
+                    emit(
+                        Resource.Error<ResultType>(
+                            apiResponse.errorMessage
+                        )
+                    )
                 }
             }
         } else {
-            emitAll(loadFromDB().map { Resource.Success(it) })
+            emitAll(loadFromDB().map {
+                Resource.Success(
+                    it
+                )
+            })
         }
     }
