+++ core/src/main/java/com/ryanrvldo/tourismapp/core/data/source/remote/network/TourismApiService.kt
@@ -1,7 +1,6 @@
 package com.ryanrvldo.tourismapp.core.data.source.remote.network
 
 import com.ryanrvldo.tourismapp.core.data.source.remote.response.ListTourismResponse
-import retrofit2.Call
 import retrofit2.http.GET
 
 interface TourismApiService {
