+++ core/src/main/java/com/ryanrvldo/tourismapp/core/data/source/remote/response/TourismResponse.kt
