+++ core/src/main/java/com/ryanrvldo/tourismapp/core/domain/model/Tourism.kt
