+++ core/src/main/res/layout/item_list_tourism.xml
