+++ settings.gradle
@@ -1,2 +1,3 @@
+include ':core'
 rootProject.name='TourismApp'
 include ':app'
